# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aimstorm Internship Registration API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registrations
    description: Internship registration operations
  - name: admin
    description: Admin authentication and management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin auth status
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /registrations:
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List all registrations
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: degree
          in: query
          schema:
            type: string
        - name: yearOfStudy
          in: query
          schema:
            type: string
        - name: timeSlot
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: canAttend
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"

    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Submit internship registration (multipart/form-data with optional resume and photo files)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegistrationBody"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /registrations/stats:
    get:
      operationId: getRegistrationStats
      tags: [registrations]
      summary: Get registration statistics
      responses:
        "200":
          description: Registration statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationStats"

  /registrations/export:
    get:
      operationId: exportRegistrations
      tags: [registrations]
      summary: Export registrations as CSV
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  /registrations/{id}:
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get a single registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateRegistration
      tags: [registrations]
      summary: Update registration status and notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegistrationBody"
      responses:
        "200":
          description: Updated registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    AdminMeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated

    CreateRegistrationBody:
      type: object
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        mobileNumber:
          type: string
        emailId:
          type: string
        address:
          type: string
        college:
          type: string
        degree:
          type: string
        department:
          type: string
        yearOfStudy:
          type: string
        university:
          type: string
        whyJoin:
          type: string
        areasOfInterest:
          type: string
        interestedInBusiness:
          type: string
        currentSkills:
          type: string
        businessKnowledge:
          type: string
        canAttendAll:
          type: string
        preferredTimeSlot:
          type: string
        agreeToProject:
          type: string
        declarationConfirmed:
          type: string
      required:
        - fullName
        - dateOfBirth
        - gender
        - mobileNumber
        - emailId
        - whyJoin
        - canAttendAll
        - agreeToProject
        - declarationConfirmed

    Registration:
      type: object
      properties:
        id:
          type: integer
        registrationNumber:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        mobileNumber:
          type: string
        emailId:
          type: string
        address:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        degree:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        yearOfStudy:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        whyJoin:
          type: string
        areasOfInterest:
          type: ["string", "null"]
        interestedInBusiness:
          type: ["string", "null"]
        currentSkills:
          type: ["string", "null"]
        businessKnowledge:
          type: ["string", "null"]
        canAttendAll:
          type: string
        preferredTimeSlot:
          type: ["string", "null"]
        agreeToProject:
          type: string
        declarationConfirmed:
          type: string
        resumePath:
          type: ["string", "null"]
        photoPath:
          type: ["string", "null"]
        status:
          type: string
        contactStatus:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        submittedAt:
          type: string
      required:
        - id
        - registrationNumber
        - fullName
        - dateOfBirth
        - gender
        - mobileNumber
        - emailId
        - whyJoin
        - canAttendAll
        - agreeToProject
        - declarationConfirmed
        - status
        - submittedAt

    UpdateRegistrationBody:
      type: object
      properties:
        status:
          type: string
        contactStatus:
          type: string
        notes:
          type: string

    RegistrationStats:
      type: object
      properties:
        total:
          type: integer
        newCount:
          type: integer
        contactedCount:
          type: integer
        selectedCount:
          type: integer
        rejectedCount:
          type: integer
        recentRegistrations:
          type: array
          items:
            $ref: "#/components/schemas/Registration"
        byDegree:
          type: array
          items:
            $ref: "#/components/schemas/StatCount"
        byTimeSlot:
          type: array
          items:
            $ref: "#/components/schemas/StatCount"
      required:
        - total
        - newCount
        - contactedCount
        - selectedCount
        - rejectedCount
        - recentRegistrations
        - byDegree
        - byTimeSlot

    StatCount:
      type: object
      properties:
        label:
          type: string
        count:
          type: integer
      required:
        - label
        - count