# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aimstorm Internship Registration API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Check admin auth status
 */
export const AdminMeResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary List all registrations
 */
export const ListRegistrationsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  degree: zod.coerce.string().optional(),
  yearOfStudy: zod.coerce.string().optional(),
  timeSlot: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
  canAttend: zod.coerce.string().optional(),
});

export const ListRegistrationsResponseItem = zod.object({
  id: zod.number(),
  registrationNumber: zod.string(),
  fullName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  mobileNumber: zod.string(),
  emailId: zod.string(),
  address: zod.string().nullish(),
  college: zod.string().nullish(),
  degree: zod.string().nullish(),
  department: zod.string().nullish(),
  yearOfStudy: zod.string().nullish(),
  university: zod.string().nullish(),
  whyJoin: zod.string(),
  areasOfInterest: zod.string().nullish(),
  interestedInBusiness: zod.string().nullish(),
  currentSkills: zod.string().nullish(),
  businessKnowledge: zod.string().nullish(),
  canAttendAll: zod.string(),
  preferredTimeSlot: zod.string().nullish(),
  agreeToProject: zod.string(),
  declarationConfirmed: zod.string(),
  resumePath: zod.string().nullish(),
  photoPath: zod.string().nullish(),
  status: zod.string(),
  contactStatus: zod.string().nullish(),
  notes: zod.string().nullish(),
  submittedAt: zod.string(),
});
export const ListRegistrationsResponse = zod.array(
  ListRegistrationsResponseItem,
);

/**
 * @summary Submit internship registration (multipart/form-data with optional resume and photo files)
 */
export const CreateRegistrationBody = zod.object({
  fullName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  mobileNumber: zod.string(),
  emailId: zod.string(),
  address: zod.string().optional(),
  college: zod.string().optional(),
  degree: zod.string().optional(),
  department: zod.string().optional(),
  yearOfStudy: zod.string().optional(),
  university: zod.string().optional(),
  whyJoin: zod.string(),
  areasOfInterest: zod.string().optional(),
  interestedInBusiness: zod.string().optional(),
  currentSkills: zod.string().optional(),
  businessKnowledge: zod.string().optional(),
  canAttendAll: zod.string(),
  preferredTimeSlot: zod.string().optional(),
  agreeToProject: zod.string(),
  declarationConfirmed: zod.string(),
});

/**
 * @summary Get registration statistics
 */
export const GetRegistrationStatsResponse = zod.object({
  total: zod.number(),
  newCount: zod.number(),
  contactedCount: zod.number(),
  selectedCount: zod.number(),
  rejectedCount: zod.number(),
  recentRegistrations: zod.array(
    zod.object({
      id: zod.number(),
      registrationNumber: zod.string(),
      fullName: zod.string(),
      dateOfBirth: zod.string(),
      gender: zod.string(),
      mobileNumber: zod.string(),
      emailId: zod.string(),
      address: zod.string().nullish(),
      college: zod.string().nullish(),
      degree: zod.string().nullish(),
      department: zod.string().nullish(),
      yearOfStudy: zod.string().nullish(),
      university: zod.string().nullish(),
      whyJoin: zod.string(),
      areasOfInterest: zod.string().nullish(),
      interestedInBusiness: zod.string().nullish(),
      currentSkills: zod.string().nullish(),
      businessKnowledge: zod.string().nullish(),
      canAttendAll: zod.string(),
      preferredTimeSlot: zod.string().nullish(),
      agreeToProject: zod.string(),
      declarationConfirmed: zod.string(),
      resumePath: zod.string().nullish(),
      photoPath: zod.string().nullish(),
      status: zod.string(),
      contactStatus: zod.string().nullish(),
      notes: zod.string().nullish(),
      submittedAt: zod.string(),
    }),
  ),
  byDegree: zod.array(
    zod.object({
      label: zod.string(),
      count: zod.number(),
    }),
  ),
  byTimeSlot: zod.array(
    zod.object({
      label: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Export registrations as CSV
 */
export const ExportRegistrationsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

/**
 * @summary Get a single registration
 */
export const GetRegistrationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRegistrationResponse = zod.object({
  id: zod.number(),
  registrationNumber: zod.string(),
  fullName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  mobileNumber: zod.string(),
  emailId: zod.string(),
  address: zod.string().nullish(),
  college: zod.string().nullish(),
  degree: zod.string().nullish(),
  department: zod.string().nullish(),
  yearOfStudy: zod.string().nullish(),
  university: zod.string().nullish(),
  whyJoin: zod.string(),
  areasOfInterest: zod.string().nullish(),
  interestedInBusiness: zod.string().nullish(),
  currentSkills: zod.string().nullish(),
  businessKnowledge: zod.string().nullish(),
  canAttendAll: zod.string(),
  preferredTimeSlot: zod.string().nullish(),
  agreeToProject: zod.string(),
  declarationConfirmed: zod.string(),
  resumePath: zod.string().nullish(),
  photoPath: zod.string().nullish(),
  status: zod.string(),
  contactStatus: zod.string().nullish(),
  notes: zod.string().nullish(),
  submittedAt: zod.string(),
});

/**
 * @summary Update registration status and notes
 */
export const UpdateRegistrationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateRegistrationBody = zod.object({
  status: zod.string().optional(),
  contactStatus: zod.string().optional(),
  notes: zod.string().optional(),
});

export const UpdateRegistrationResponse = zod.object({
  id: zod.number(),
  registrationNumber: zod.string(),
  fullName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  mobileNumber: zod.string(),
  emailId: zod.string(),
  address: zod.string().nullish(),
  college: zod.string().nullish(),
  degree: zod.string().nullish(),
  department: zod.string().nullish(),
  yearOfStudy: zod.string().nullish(),
  university: zod.string().nullish(),
  whyJoin: zod.string(),
  areasOfInterest: zod.string().nullish(),
  interestedInBusiness: zod.string().nullish(),
  currentSkills: zod.string().nullish(),
  businessKnowledge: zod.string().nullish(),
  canAttendAll: zod.string(),
  preferredTimeSlot: zod.string().nullish(),
  agreeToProject: zod.string(),
  declarationConfirmed: zod.string(),
  resumePath: zod.string().nullish(),
  photoPath: zod.string().nullish(),
  status: zod.string(),
  contactStatus: zod.string().nullish(),
  notes: zod.string().nullish(),
  submittedAt: zod.string(),
});